+++ src/main/java/org/ideplugins/ci_pipeline_lint/completion/GitlabCIVariablesCompletionContributor.java
@@ -3,6 +3,7 @@ package org.ideplugins.ci_pipeline_lint.completion;
 import com.intellij.codeInsight.completion.CompletionContributor;
 import com.intellij.codeInsight.completion.CompletionType;
 import com.intellij.patterns.PlatformPatterns;
+import com.intellij.patterns.StandardPatterns;
 import org.jetbrains.yaml.YAMLElementTypes;
 import org.jetbrains.yaml.YAMLLanguage;
 import org.jetbrains.yaml.psi.YAMLSequenceItem;
@@ -14,14 +15,22 @@ public class GitlabCIVariablesCompletionContributor extends CompletionContributo
     public GitlabCIVariablesCompletionContributor() {
         extend(CompletionType.BASIC,
                 PlatformPatterns.psiElement().withLanguage(YAMLLanguage.INSTANCE)
-                        .inFile(PlatformPatterns.psiFile().withName(".gitlab-ci.yml")),
+                        .inFile(PlatformPatterns.psiFile().withName(".gitlab-ci.yml", ".gitlab-ci.yaml")),
+                new GitLabCompletionProvider()
+        );
+
+        extend(CompletionType.BASIC,
+                PlatformPatterns.psiElement().withLanguage(YAMLLanguage.INSTANCE)
+                        .inFile(PlatformPatterns.psiFile().withName(
+                                StandardPatterns.string().matches("(\\.)?gitlab-ci-.*\\.(yml|yaml)")
+                        )),
                 new GitLabCompletionProvider()
         );
 
         extend(CompletionType.BASIC,
                 PlatformPatterns.psiElement().withLanguage(YAMLLanguage.INSTANCE)
                         .inFile(PlatformPatterns.psiFile().withName(".gitlab-ci.yml"))
-                                .and(psiElement().withElementType(YAMLElementTypes.TEXT_SCALAR_ITEMS)
+                        .and(psiElement().withElementType(YAMLElementTypes.TEXT_SCALAR_ITEMS)
                                 .and(psiElement().withText("$"))
                                 .and(psiElement().withParent(psiElement(YAMLElementTypes.SCALAR_PLAIN_VALUE)))
                                 .and(psiElement().withSuperParent(2, psiElement(YAMLSequenceItem.class))))