+++ src/main/java/org/ideplugins/ci_pipeline_lint/icons/Icons.java
@@ -1,0 +1,10 @@
+package org.ideplugins.ci_pipeline_lint.icons;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+public interface Icons {
+    Icon FileType = IconLoader.getIcon("/icons/file-lines.svg", Icons.class);
+    Icon gitlabProjectIdScreenshot = IconLoader.getIcon("/icons/gitlabProjectID.png", Icons.class);
+}